extending obs builder to handle rarer cases
