+++ docs/uml/BUFR_ObsBuilder.puml
@@ -16,14 +16,13 @@ class Logger
 
 abstract class ObsBuilder
 {
-  - input_path : str
-  - mapping_path : str
+  - mapping_path : dict
   - log : Logger
 
-  + create_obs_group(input_path, mapping_path, category, env)
-  + create_obs_file(input_path, mapping_path, type="netcdf", output_path)
   + make_obs(comm)
   # _make_description()
+  + create_obs_group(input_path, mapping_path, category, env)
+  + create_obs_file(input_path, mapping_path, type="netcdf", output_path)
 }
 
 class MarineObsBuilder
