+++ python/bufr/obs_builder/__init__.py
@@ -1,5 +1,5 @@
 from .logger import Logger
 
+from .add_main_functions import add_main_functions
 from .obs_builder import ObsBuilder
 from .obs_builder import add_encoder_type
-from .obs_builder import add_main_functions
